Move calculator data type out of bolus package hierarchy
