+++ data/factory/standard_test.go
@@ -18,10 +18,10 @@ import (
 	"github.com/tidepool-org/platform/data/types/base/blood/glucose/continuous"
 	"github.com/tidepool-org/platform/data/types/base/blood/glucose/selfmonitored"
 	"github.com/tidepool-org/platform/data/types/base/blood/ketone"
-	"github.com/tidepool-org/platform/data/types/base/bolus/calculator"
 	"github.com/tidepool-org/platform/data/types/base/bolus/combination"
 	"github.com/tidepool-org/platform/data/types/base/bolus/extended"
 	"github.com/tidepool-org/platform/data/types/base/bolus/normal"
+	"github.com/tidepool-org/platform/data/types/base/calculator"
 	"github.com/tidepool-org/platform/data/types/base/device/alarm"
 	"github.com/tidepool-org/platform/data/types/base/device/calibration"
 	"github.com/tidepool-org/platform/data/types/base/device/prime"
